SMASHING
- Added in Smashing mechanic. I didn't do it in a good way to be able to add more actions easily, but it's fine for now.

- Box obstacle is now wider so that you have to SMASH it early on.

+++ BasicSideScroller/Assets/Scripts/GameManagement/GameManager.cs
@@ -51,7 +51,7 @@ public class GameManager : MonoBehaviour
     void Update()
     {
         gameTime += Time.deltaTime;
-        GameTimerText.text = gameTime.ToString();
+        GameTimerText.text = ((int)gameTime).ToString();
 /* 
         spawnTimer += Time.deltaTime;
         if (spawnTimer >= timeToSpawn)
